upd: Wallpaper tab & gui.setup.py
