+++ walset/gui/style.css
@@ -1,13 +1,52 @@
-/* walset/gui/style.css */
-
-.preference-button {
-    background-color: transparent;
+/* Buttons */
+.moreoptions-button {
+	box-shadow: 0px 0px 0px 0px transparent;
+	border-color: transparent;
     background-image: none;
+	border: 0px;
 }
 
-.preference-button:hover {
+.moreoptions-button:hover {
 	opacity: 0.7;
 	border-color: transparent;
 	background-color: transparent;
 	box-shadow: 0px 0px 0px 0px transparent;
 }
+
+.moreoptions-button:active {
+	border-color: transparent;
+	box-shadow: 0px 0px 0px 0px transparent;
+}
+
+.start-buttons-labels {
+	padding-left: 6px;
+	padding-right: 6px;	
+}
+
+.start-buttons-labels:hover, .start-buttons-labels:active .start-buttons-labels:checked {
+	padding-left: 12px;
+	padding-right: 12px;
+}
+
+.color-picker {
+	border-radius: 0px;
+	border: 0px;
+	padding: 0px;
+	margin: 0px
+}
+
+.wallpaper-preview {
+	background-color: alpha(@theme_fg_color, 0.15);
+}
+
+/*
+ 
+==== Gtk 3 Color Legend ====
+ 
+@theme_bg_color:	The base background for windows and dialogs.
+@theme_fg_color:	The main text color (usually dark gray/black).
+@theme_base_color:	The background for text entries and lists (usually pure white).
+@theme_text_color:	The text color for entries and lists.
+@theme_selected_bg_color:	The "Accent" color (blue by default) for active tabs or selections.
+@theme_selected_fg_color:
+*/
